move SQLiteDB creation file
Moved it from resources to util because it makes sense to have a util because that is exactly what it is
